Add files via upload
For this project, I have created a quiz game with a point system that gives you the correct answer whenever you get a question wrong and also when your input is not one of the multiple choices it will asks for your input again. Once you are completed with the game it would ask you if you want to play again or not. This project was mostly to help me learn how functions works with each other and also how 2d list work within coding.